IDNA fast path for ASCII (#3)
We can skip most of the domain processing if the input is a valid ASCII string.
This PR also removes any reduntant heap allocation for ASCII, so the only thing heap allocated in HostParser.Parse for a regular domain is a result itself.

top100.txt benchmark: 39% less memory allocations and 32% less execution time